+++ CHANGELOG.md
@@ -4,6 +4,16 @@ The format of this changelog is based on
 [Keep a Changelog](https://keepachangelog.com/), and this project adheres to
 [Semantic Versioning](https://semver.org/).
 
+## Upcoming
+
+  - Composite components can define `_build_subcomponents` to return a `NamedTuple` with keys that differ from component names
+
+### Fixed
+
+  - Rounding no longer fails when radius is less than `min_side_len` only due to numerical precision issues
+  - Circular arcs in rounded polygons will no longer occasionally produce very short edges near the endpoints, and are instead now drawn with equally spaced points including the endpoints
+  - `Turn` segments with `SimpleTrace` or `SimpleCPW` styles now use `atol` to determine the discretization; this is faster and in some cases more accurate than the fallback method using `adapted_grid`
+
 ## 1.1.1 (2025-04-16)
 
   - Improved performance of nested `CompositeComponent`s by storing hooks after first computation